requirements.txt
amqp==1.4.9
anyio==3.6.2
anyjson==0.3.3
anytree==2.9.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
billiard==3.3.0.23
bleach==6.0.0
blessed==1.20.0
bs4==0.0.1
celery==3.1.26.post2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
comm==0.1.2
contourpy==1.0.6
cron-descriptor==1.3.0
cryptography==38.0.3
cx-Oracle==8.3.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20230430
Django==5.0.3
django-admin-csvexport==2.2
django-admin-filter==0.7
django-admin-filters==0.1.3
django-admin-list-filters==1.1
django-appconf==1.0.6
django-autoslug==1.9.9
django-celery==3.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cities-light==3.10.1
django-cookie-consent==0.4.0
django-cors-headers==3.13.0
django-countries==7.5.1
django-crispy-forms==2.1
django-debug-toolbar==4.0.0
django-elasticsearch-dsl==7.2.2
django-environ==0.10.0
django-extensions==3.2.3
django-filter==22.1
django-import-export==3.2.0
django-microsoft-authentication==0.1.5
django-modeltree==0.5
django-multi-form-view==2.0.1
django-picklefield==3.1
django-q==1.3.9
django-ranged-response==0.2.0
django-rangefilter==0.9.1
django-redis==5.2.0
django-simple-captcha==0.6.0
django-storages==1.13.1
django-tastypie==0.14.4
django-timezone-field==5.0
django_csp==3.8
djangorestframework==3.14.0
elastic-transport==8.4.0
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
fpdf2==2.7.5
fqdn==1.5.1
gunicorn==20.1.0
html5lib==1.1
htmlmerger==0.1.46
httpie==3.2.2
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kombu==3.0.37
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
msal==1.22.0
multidict==6.0.5
mysql-connector-python==8.0.31
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pdf2image==1.16.3
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip-chill==1.0.1
platformdirs==2.6.2
polars==0.20.15
progressbar2==4.3.2
prometheus-client==0.16.0
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.14.0
PyJWT==2.6.0
pymemcache==4.0.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pysftp==0.2.9
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-memcached==1.59
python-mimeparse==1.6.0
python-utils==3.8.1
pytz==2022.6
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
redis==2.10.6
requests==2.28.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
Send2Trash==1.8.0
sentry-sdk==1.11.0
six==1.16.0
smile-id-core==2.1.5
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.5.0
tabulate==0.9.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.8.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.8
uri-template==1.2.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
whitenoise==6.4.0
widgetsnbextension==4.0.5
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
